Refactor project structure and update database schema

+++ RR.AI-Chat/RR.AI-Chat.Dto/ChatCompletionDto.cs
@@ -6,10 +6,10 @@
 
         public string? Message { get; set; }
 
-        public int? InputTokenCount { get; set; }
+        public long? InputTokenCount { get; set; }
 
-        public int? OutputTokenCount { get; set; }
+        public long? OutputTokenCount { get; set; }
 
-        public int? TotalTokenCount { get; set; }
+        public long? TotalTokenCount { get; set; }
     }
 }
